Refactor into separate modules
This commit moves the prefix building functions to a separate module,
and new_tenant, drop_tenant to a different module.

The public api of the `Apartmentex` module remains the same, since
`drop_tenant` and `new_tenant` are delegated from the top level module
to the `Apartmentex.TenantActions` module.

Now all of the code in `lib/apartmentex.ex` deals with building queries
to pass to Ecto.
