+++ lib/apartmentex/prefix_builder.ex
@@ -1,0 +1,15 @@
+defmodule Apartmentex.PrefixBuilder do
+  @schema_prefix Application.get_env(:apartmentex, :schema_prefix) || "tenant_"
+
+  def build_prefix(tenant) when is_integer(tenant) do
+    @schema_prefix <> Integer.to_string(tenant)
+  end
+
+  def build_prefix(tenant) when is_binary(tenant) do
+    @schema_prefix <> tenant
+  end
+
+  def build_prefix(tenant) do
+    @schema_prefix <> Integer.to_string(tenant.id)
+  end
+end
